+++ app/costasiella/schema/organization_subscription.py
@@ -63,7 +63,7 @@ def validate_create_update_input(input, update=False):
 class OrganizationSubscriptionNodeInterface(graphene.Interface):
     id = graphene.GlobalID()
     subscription_unit_display = graphene.String()
-    price_today = graphene.Float()
+    price_today = graphene.Decimal()
     price_today_display = graphene.String()
 
 
@@ -73,16 +73,13 @@ class OrganizationSubscriptionNode(DjangoObjectType):
         filter_fields = ['archived']
         interfaces = (graphene.relay.Node, OrganizationSubscriptionNodeInterface)
 
-
     def resolve_subscription_unit_display(self, info):
         return display_subscription_unit(self.subscription_unit)
-
 
     def resolve_price_today_display(self, info):
         today = timezone.now().date()
 
         return self.get_price_on_date(today, display=True)
-
 
     def resolve_price_today(self, info):
         today = timezone.now().date()
@@ -112,7 +109,7 @@ class OrganizationSubscriptionQuery(graphene.ObjectType):
             return OrganizationSubscription.objects.filter(archived = archived).order_by('name')
 
         # Return only public non-archived locations
-        return OrganizationSubscription.objects.filter(display_public = True, archived = False).order_by('name')
+        return OrganizationSubscription.objects.filter(display_public=True, archived=False).order_by('name')
 
 
 class CreateOrganizationSubscription(graphene.relay.ClientIDMutation):
@@ -129,9 +126,9 @@ class CreateOrganizationSubscription(graphene.relay.ClientIDMutation):
         credit_validity = graphene.Int(required=False, default_value=0)
         unlimited = graphene.Boolean(required=True, default_value=False)
         terms_and_conditions = graphene.String(required=False, default_value="")
-        registration_fee = graphene.Float(required=False, default_value=0)
+        registration_fee = graphene.Decimal(required=False, default_value=0)
         organization_membership = graphene.ID(required=False, default_value="")
-        quick_stats_amount = graphene.Float(required=False, default_value=0)
+        quick_stats_amount = graphene.Decimal(required=False, default_value=0)
         finance_glaccount = graphene.ID(required=False, default_value="")
         finance_costcenter = graphene.ID(required=False, default_value="")      
         
@@ -192,9 +189,9 @@ class UpdateOrganizationSubscription(graphene.relay.ClientIDMutation):
         credit_validity = graphene.Int(required=False, default_value=0)
         unlimited = graphene.Boolean(required=True, default_value=False)
         terms_and_conditions = graphene.String(required=False, default_value="")
-        registration_fee = graphene.Float(required=False, default_value=0)
+        registration_fee = graphene.Decimal(required=False, default_value=0)
         organization_membership = graphene.ID(required=False, default_value="")
-        quick_stats_amount = graphene.Float(required=False, default_value=0)
+        quick_stats_amount = graphene.Decimal(required=False, default_value=0)
         finance_glaccount = graphene.ID(required=False, default_value="")
         finance_costcenter = graphene.ID(required=False, default_value="")
